wip - refactor remove abstract klass
