--- engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/sink/BaseKafkaAvroSinkFactory.scala
@@ -1,48 +1,0 @@
-package pl.touk.nussknacker.engine.avro.sink
-
-import cats.data.Validated
-import org.apache.avro.Schema
-import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{CustomNodeError, NodeId}
-import pl.touk.nussknacker.engine.api.process.SinkFactory
-import pl.touk.nussknacker.engine.api.typed.CustomNodeValidationException
-import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
-import pl.touk.nussknacker.engine.api.{LazyParameter, MetaData}
-import pl.touk.nussknacker.engine.avro.encode.{OutputValidator, ValidationMode}
-import pl.touk.nussknacker.engine.avro.schemaregistry.SchemaVersionOption
-import pl.touk.nussknacker.engine.avro.serialization.KafkaAvroSerializationSchemaFactory
-import pl.touk.nussknacker.engine.avro.{AvroSchemaDeterminer, SchemaDeterminerErrorHandler}
-import pl.touk.nussknacker.engine.flink.api.process.FlinkSink
-import pl.touk.nussknacker.engine.kafka.{KafkaConfig, PreparedKafkaTopic}
-
-abstract class BaseKafkaAvroSinkFactory extends SinkFactory {
-
-  override def requiresOutput: Boolean = false
-
-  protected def createSink(preparedTopic: PreparedKafkaTopic,
-                           version: SchemaVersionOption,
-                           key: LazyParameter[AnyRef],
-                           valueEither: Either[List[(String, LazyParameter[AnyRef])], LazyParameter[AnyRef]],
-                           kafkaConfig: KafkaConfig,
-                           serializationSchemaFactory: KafkaAvroSerializationSchemaFactory,
-                           schemaDeterminer: AvroSchemaDeterminer, validationMode: ValidationMode)
-                          (implicit processMetaData: MetaData,
-                           nodeId: NodeId): FlinkSink = {
-    //This is a bit redundant, since we already validate during creation
-    val schemaData = schemaDeterminer.determineSchemaUsedInTyping.valueOr(SchemaDeterminerErrorHandler.handleSchemaRegistryErrorAndThrowException)
-    // TODO
-//    validateValueType(value.returnType, schemaData.schema, validationMode).valueOr(err => throw new CustomNodeValidationException(err.message, err.paramName, null))
-    val schemaUsedInRuntime = schemaDeterminer.toRuntimeSchema(schemaData)
-
-    val clientId = s"${processMetaData.id}-${preparedTopic.prepared}"
-    new KafkaAvroSink(preparedTopic, version, key, valueEither, kafkaConfig, serializationSchemaFactory,
-      schemaData.serializableSchema, schemaUsedInRuntime.map(_.serializableSchema), clientId, validationMode)
-  }
-
-  /**
-    * Currently we check only required fields, because our typing mechanism doesn't support optionally fields
-    */
-  protected def validateValueType(valueType: TypingResult, schema: Schema, validationMode: ValidationMode)(implicit nodeId: NodeId): Validated[CustomNodeError, Unit] = {
-    OutputValidator.validateOutput(valueType, schema, validationMode)
-  }
-
-}
